fix: Correct comparison for last message
The batch_index was Some(-1) or None for the same message depending on
whether we got the message id from Reader::get_last_message_id() or from
the current message's message_id().

To address this, we wait for a timeout and then only check the ledger
and entry id as well as the partition to see if we reached the end of
the topic.